+++ db/migrate/20150521115934_add_indexes_to_subject_events.rb
@@ -1,0 +1,6 @@
+class AddIndexesToSubjectEvents < ActiveRecord::Migration
+  def change
+    add_index :subject_events, :subject_id
+    add_index :subject_events, :event_id
+  end
+end